requirements.txt
opencv-python
pyserial
numpy
dlib